upload/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
gspread
google-auth